Fixing compiler warnings, part 2

+++ Cantera/src/base/vec_functions.h
@@ -175,7 +175,7 @@ namespace Cantera {
    *  @param len  Length of the copy
    */
   inline void fbo_copy_dbl_1(doublereal * const copyTo, const doublereal * const copyFrom, 
-			     const int len) {
+			     const size_t len) {
     if (len > 0) {
       (void) memcpy((void *)copyTo, (const void *)copyFrom, len * sizeof(doublereal));
     }
@@ -190,7 +190,7 @@ namespace Cantera {
    *  @param len  Length of the copy
    */
   inline void fvo_copy_dbl_1(std::vector<doublereal> &copyTo, const std::vector<doublereal> &copyFrom, 
-			     const int len) {
+			     const size_t len) {
     if (len > 0) {
       (void) memcpy((void *)(&copyTo[0]), (const void *)(&copyFrom[0]), len * sizeof(doublereal));
     }
@@ -203,7 +203,7 @@ namespace Cantera {
    *  @param v    Vector to be zeroed
    *  @param len  Length of the copy
    */
-  inline void fbo_zero_dbl_1(doublereal * const v, const int len) {
+  inline void fbo_zero_dbl_1(doublereal * const v, const size_t len) {
     if (len > 0) {
       (void) memset((void *)v, 0, len * sizeof(doublereal));
     }
@@ -216,7 +216,7 @@ namespace Cantera {
    *  @param v    Vector to be zeroed
    *  @param len  Length of the copy
    */
-  inline void fvo_zero_dbl_1(std::vector<doublereal> &v, const int len) {
+  inline void fvo_zero_dbl_1(std::vector<doublereal> &v, const size_t len) {
     if (len > 0) {
       (void) memset((void *)(&v[0]), 0, len * sizeof(doublereal));
     }
