+++ Cantera/src/thermo/Nasa9PolyMultiTempRegion.cpp
@@ -107,7 +107,7 @@ namespace Cantera {
     m_highT = m_regionPts[m_numTempRegions-1]->maxTemp();
     m_Pref = m_regionPts[0]->refPressure();
     m_index = m_regionPts[0]->speciesIndex();
-    for (int i = 0; i < m_numTempRegions; i++) {
+    for (size_t i = 0; i < m_numTempRegions; i++) {
       m_lowerTempBounds[i] = m_regionPts[i]->minTemp();
       if (m_regionPts[i]->speciesIndex() != m_index) {
 	throw CanteraError("Nasa9PolyMultiTempRegion::Nasa9PolyMultiTempRegion",
@@ -145,7 +145,7 @@ namespace Cantera {
     m_currRegion(b.m_currRegion)
   {
     m_regionPts.resize(m_numTempRegions);
-    for (int i = 0; i < m_numTempRegions; i++) {
+    for (size_t i = 0; i < m_numTempRegions; i++) {
       Nasa9Poly1 * dptr = b.m_regionPts[i];
       m_regionPts[i] = new Nasa9Poly1(*dptr);
     }
@@ -158,7 +158,7 @@ namespace Cantera {
   Nasa9PolyMultiTempRegion& 
   Nasa9PolyMultiTempRegion::operator=(const Nasa9PolyMultiTempRegion& b) {
     if (&b != this) {
-      for (int i = 0; i < m_numTempRegions; i++) {
+      for (size_t i = 0; i < m_numTempRegions; i++) {
 	delete m_regionPts[i];
 	m_regionPts[i] = 0;
       }
@@ -170,7 +170,7 @@ namespace Cantera {
       m_lowerTempBounds = b.m_lowerTempBounds;
       m_currRegion = b.m_currRegion;
       m_regionPts.resize(m_numTempRegions);
-      for (int i = 0; i < m_numTempRegions; i++) {
+      for (size_t i = 0; i < m_numTempRegions; i++) {
 	m_regionPts[i] = new Nasa9Poly1(*(b.m_regionPts[i]));
       }
     }
@@ -179,7 +179,7 @@ namespace Cantera {
 
   // Destructor
   Nasa9PolyMultiTempRegion::~Nasa9PolyMultiTempRegion() {
-    for (int i = 0; i < m_numTempRegions; i++) {
+    for (size_t i = 0; i < m_numTempRegions; i++) {
       delete m_regionPts[i];
       m_regionPts[i] = 0;
     }
@@ -308,7 +308,7 @@ namespace Cantera {
     tPoly[6]  = std::log(temp);
     // Now find the region
     m_currRegion = 0;
-    for (int i = 1; i < m_numTempRegions; i++) {
+    for (size_t i = 1; i < m_numTempRegions; i++) {
       if (temp < m_lowerTempBounds[i]) {
 	break;
       }
@@ -347,7 +347,7 @@ namespace Cantera {
     int n_tmp = 0;;
     int type_tmp = 0;
     double pref_tmp = 0.0;
-    for (int iReg = 0; iReg < m_numTempRegions; iReg++) {
+    for (size_t iReg = 0; iReg < m_numTempRegions; iReg++) {
       m_regionPts[iReg]->reportParameters(n_tmp, type_tmp,
 					  coeffs[index], coeffs[index+1],
 					  pref_tmp, ctmp);
@@ -366,12 +366,10 @@ namespace Cantera {
    */
   void Nasa9PolyMultiTempRegion::modifyParameters(doublereal* coeffs) {
     int index = 3;
-    for (int iReg = 0; iReg < m_numTempRegions; iReg++) {
+    for (size_t iReg = 0; iReg < m_numTempRegions; iReg++) {
       m_regionPts[iReg]->modifyParameters(coeffs + index);
       index += 11;
     }   
   }
 
-
 }
-
