+++ Cantera/src/thermo/Phase.h
@@ -220,7 +220,7 @@ namespace Cantera {
      * The index is used in the Python and matlab interfaces to 
      * index into a list of ThermoPhase objects
      */
-    int index() const;
+    size_t index() const;
 
     //! Sets the index of the phase
     /*!
@@ -229,7 +229,7 @@ namespace Cantera {
      *
      * @param m Integer index of the phase
      */
-    void setIndex(int m);
+    void setIndex(size_t m);
 
     //! Save the current internal state of the phase
     /*!
@@ -434,7 +434,7 @@ namespace Cantera {
      *
      * @return Mole fraction of the species
      */
-    doublereal moleFraction(int k) const;
+    doublereal moleFraction(size_t k) const;
         
     //! Return the mole fraction of a single species
     /*!
@@ -450,7 +450,7 @@ namespace Cantera {
      *
      * @return Mass Fraction of the species
      */
-    doublereal massFraction(int k) const;
+    doublereal massFraction(size_t k) const;
 
     //! Return the mass fraction of a single species
     /*!
@@ -466,7 +466,7 @@ namespace Cantera {
     doublereal chargeDensity() const;
 
     /// Returns the number of spatial dimensions (1, 2, or 3)
-    int nDim() const {return m_ndim;}
+    size_t nDim() const {return m_ndim;}
 
     //! Set the number of spatial dimensions (1, 2, or 3)
     /*!
@@ -475,7 +475,7 @@ namespace Cantera {
      *
      * @param ndim   Input number of dimensions.
      */
-    void setNDim(int ndim) {m_ndim = ndim;}
+    void setNDim(size_t ndim) {m_ndim = ndim;}
 
     /** 
      *  Finished adding species, prepare to use them for calculation
@@ -495,19 +495,19 @@ namespace Cantera {
      * to restate it in here, so that the declarations in the two
      * base classes become hidden.
      */
-    int m_kk;
+    size_t m_kk;
 
     /**
      * m_ndim is the dimensionality of the phase.  Volumetric
      * phases have dimensionality 3 and surface phases have
      * dimensionality 2.
      */
-    int m_ndim;
+    size_t m_ndim;
     /**
      * m_index is the index of the phase
      * 
      */
-    int m_index;
+    size_t m_index;
 
   private:
